requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
streamlit>=1.37.0
requests>=2.32.0
agno>=1.0.0
python-dotenv>=1.0.1
